+++ app/javascript/lib/shared/index.js
@@ -12,6 +12,7 @@ import "./modules/tabs.js";
 import "./modules/velocity_settings.js";
 import "./modules/shareContent.js";
 import "./modules/separate-tabs.js";
+import { createScaleColors } from "./modules/createScaleColors.js";
 import accounting from "accounting";
 import { SETTINGS } from "./modules/accounting_settings.js";
 accounting.settings = SETTINGS;
@@ -39,6 +40,7 @@ import { Middleware } from "./modules/middleware.js";
 export {
   AUTOCOMPLETE_DEFAULTS,
   accounting,
+  createScaleColors,
   d3locale,
   isDesktop,
   isMobile,
