+++ app/javascript/lib/shared/modules/createScaleColors.js
@@ -1,0 +1,11 @@
+import { scaleOrdinal } from 'd3-scale';
+
+export function createScaleColors(values, arrayDomain) {
+  let colorsGobiertoExtend = ["#12365b", "#118e9c", "#ff766c", "#f7b200", "#158a2c", "#94d2cf", "#3a78c3", "#15dec5", "#6a7f2f", "#55f17b"]
+  colorsGobiertoExtend = colorsGobiertoExtend.slice(0, values)
+
+  const colors = scaleOrdinal()
+    .domain(arrayDomain)
+    .range(colorsGobiertoExtend);
+  return colors;
+}
